[refactor] 가게 등록 요청 DTO 패키지 이동
